add validate method
One could check for the code of call()'s response. But that could be misleading.
If an validation error occurs, a transaction object is build with the code 400.
But the HTTP method could be wrong. E.g. the called operation requires a POST
request, but a validation error occurs, the http method of the transaction is GET.

Furthermore it's not clear that the request isn't sent and the response is built by the module

Having those validation methods it's easy to check the parameters before call() is called.